+++ TinkoffASDKUI/TinkoffASDKUI/PaymentModule/Payments/YandexPayPaymentProcess.swift
@@ -8,11 +8,23 @@
 import Foundation
 import TinkoffASDKCore
 
-final class YandexPayPaymentProcess {
+final class YandexPayPaymentProcess: PaymentProcess {
+    // MARK: PaymentProcess properties
+
+    let paymentFlow: PaymentFlow
+    let paymentSource: PaymentSourceData
+
+    var paymentId: String? {
+        switch paymentFlow {
+        case .full:
+            return _paymentId.wrappedValue
+        case let .finish(paymentId, _):
+            return paymentId
+        }
+    }
+
     // MARK: Dependencies
 
-    private let paymentOptions: PaymentOptions
-    private let base64Token: String
     private let paymentService: IAcquiringPaymentsService
     private let threeDSService: IAcquiringThreeDSService
     private let threeDSDeviceInfoProvider: IThreeDSDeviceInfoProvider
@@ -27,39 +39,30 @@ final class YandexPayPaymentProcess {
     // MARK: Init
 
     init(
-        paymentOptions: PaymentOptions,
-        base64Token: String,
+        paymentFlow: PaymentFlow,
+        paymentSource: PaymentSourceData,
         paymentService: IAcquiringPaymentsService,
         threeDSService: IAcquiringThreeDSService,
         threeDSDeviceInfoProvider: IThreeDSDeviceInfoProvider,
         delegate: PaymentProcessDelegate
     ) {
-        self.paymentOptions = paymentOptions
-        self.base64Token = base64Token
+        self.paymentFlow = paymentFlow
+        self.paymentSource = paymentSource
         self.paymentService = paymentService
         self.threeDSService = threeDSService
         self.threeDSDeviceInfoProvider = threeDSDeviceInfoProvider
         self.delegate = delegate
     }
-}
 
-// MARK: - PaymentProcess
-
-extension YandexPayPaymentProcess: PaymentProcess {
-    var paymentId: String? {
-        _paymentId.wrappedValue
-    }
-
-    var paymentFlow: PaymentFlow {
-        .full(paymentOptions: paymentOptions)
-    }
-
-    var paymentSource: TinkoffASDKCore.PaymentSourceData {
-        .yandexPay(base64Token: base64Token)
-    }
+    // MARK: PaymentProcess
 
     func start() {
-        initPayment(data: .data(with: paymentOptions))
+        switch paymentFlow {
+        case let .full(paymentOptions):
+            initPayment(data: .data(with: paymentOptions))
+        case let .finish(paymentId, _):
+            finishAuthorize(data: createFinishAuthorizeData(paymentId: paymentId))
+        }
     }
 
     func cancel() {
@@ -71,14 +74,14 @@ extension YandexPayPaymentProcess: PaymentProcess {
 // MARK: - Helpers
 
 private extension YandexPayPaymentProcess {
-    func initPayment(data: PaymentInitData) {
+    private func initPayment(data: PaymentInitData) {
         let request = paymentService.initPayment(data: data) { [weak self] result in
             guard let self = self, !self.isCancelled.wrappedValue else { return }
 
             switch result {
             case let .success(initPayload):
                 self._paymentId.store(newValue: initPayload.paymentId)
-                self.finishAuthorize(initPayload: initPayload)
+                self.finishAuthorize(data: self.createFinishAuthorizeData(paymentId: initPayload.paymentId))
             case let .failure(error):
                 self.delegate?.paymentDidFailed(self, with: error, cardId: nil, rebillId: nil)
             }
@@ -87,15 +90,8 @@ private extension YandexPayPaymentProcess {
         currentRequest.store(newValue: request)
     }
 
-    private func finishAuthorize(initPayload: InitPayload) {
-        let finishData = FinishAuthorizeData(
-            paymentId: initPayload.paymentId,
-            paymentSource: paymentSource,
-            infoEmail: paymentOptions.customerOptions?.email,
-            deviceInfo: threeDSDeviceInfoProvider.createDeviceInfo(threeDSCompInd: .threeDSCompInd)
-        )
-
-        let request = paymentService.finishAuthorize(data: finishData) { [weak self] result in
+    private func finishAuthorize(data: FinishAuthorizeData) {
+        let request = paymentService.finishAuthorize(data: data) { [weak self] result in
             guard let self = self, !self.isCancelled.wrappedValue else { return }
 
             switch result {
@@ -107,6 +103,15 @@ private extension YandexPayPaymentProcess {
         }
 
         currentRequest.store(newValue: request)
+    }
+
+    private func createFinishAuthorizeData(paymentId: String) -> FinishAuthorizeData {
+        FinishAuthorizeData(
+            paymentId: paymentId,
+            paymentSource: paymentSource,
+            infoEmail: paymentFlow.customerOptions?.email,
+            deviceInfo: threeDSDeviceInfoProvider.createDeviceInfo(threeDSCompInd: .threeDSCompInd)
+        )
     }
 
     private func handleFinishAuthorize(payload: FinishAuthorizePayload) {