+++ TinkoffASDKUI/TinkoffASDKUI/YandexPay/Interfaces/IYandexPayPaymentFlow.swift
@@ -9,10 +9,6 @@ import Foundation
 
 /// Объект, получающий результат работы `YandexPay` и отображающий UI при проведении платежа
 public protocol IYandexPayPaymentFlow: AnyObject {
-    /// Объект, получающий уведомление о завершении оплаты. Удерживается слабой ссылкой
-    var output: IYandexPayPaymentFlowOutput? { get set }
-    /// Объект, через которой запрашивается `UIViewController` для отображения UI при проведении платежа
-    var presentingViewControllerProvider: IPresentingViewControllerProvider? { get set }
     /// Начинает проведение оплаты с отображением соответствующего UI
-    func start(with paymentOption: PaymentOptions, base64Token: String)
+    func start(with paymentFlow: PaymentFlow, base64Token: String)
 }